few channges into export workflow
