+++ ipl/database/scripts/select_images.sql
@@ -1,2 +1,1 @@
-select image_id, field_id, data, image_data
-from image where field_id = ? and data between ? and ?'
+select * from image where field_id = ? and capture_data between ? and ?
