+++ ipl/errors.py
@@ -1,6 +1,6 @@
-import rasterio
 import warnings
 
+import rasterio
 
 # warnings.filterwarnings("ignore", category=RuntimeWarning)
 # warnings.filterwarnings("ignore", category=rasterio.errors.NotGeoreferencedWarning)
