+++ requirements.txt
@@ -1,5 +1,6 @@
-pandas
-scipy==1.4.1
 matplotlib==3.2.0
 rasterio==1.0.24
 numpy==1.18.1
+tabulate==0.8.6
+pandas==1.0.1
+scipy==1.4.1
